Conditional python package removal

+++ jenkins/internal-ci/centos-7.8.2003/main/cortx-prereq.groovy
@@ -98,7 +98,9 @@ pipeline {
 				sh label: 'Clean-up', script: '''
 				set +x
 				rm -rf /etc/yum.repos.d/cortx-storage.colo.seagate.com* /etc/pip.conf /root/rpmbuild/RPMS/x86_64/*.rpm
-                pip3 uninstall -r /opt/seagate/cortx/python-deps/python-requirements.txt -y
+                if [ -f /opt/seagate/cortx/python-deps/python-requirements.txt ]; then
+                    pip3 uninstall -r /opt/seagate/cortx/python-deps/python-requirements.txt -y
+                fi
 				yum erase cortx-prereq -y
 				'''
 		}
